+++ src/index-umd.js
@@ -1,3 +1,0 @@
-import StreetView from './ol-street-view';
-
-export default StreetView;